refactor: scheduled tasks

+++ src/presentation/service/marketplace.go
@@ -121,7 +121,7 @@ func (service *MarketplaceService) InstallCatalogItem(
 		taskCmd, _ := valueObject.NewUnixCommand(cliCmd)
 		taskTag, _ := valueObject.NewScheduledTaskTag("marketplace")
 		taskTags := []valueObject.ScheduledTaskTag{taskTag}
-		timeoutSeconds := uint(600)
+		timeoutSeconds := uint16(600)
 
 		scheduledTaskCreateDto := dto.NewCreateScheduledTask(
 			taskName, taskCmd, taskTags, &timeoutSeconds, nil,
